requirements.txt
streamlit>=1.55.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
crewai>=1.12.0
crewai-tools>=1.12.0
yfinance>=1.2.0
requests>=2.31.0
scipy>=1.12.0
plotly>=6.6.0
openpyxl>=3.1.0
mftool>=3.0
